updated the test helper to not emit warnings when compiling

+++ test/test_helper.h
@@ -1,4 +1,4 @@
 #include <assert.h>
 void * _l, * _r;
-#define assertEqual(l, r) if ((_l = l) != (_r = r)) { printf("Assertion Failed: assertEqual(%s, %s)\n    %llu != %llu (%p != %p)\n# %s:%d\n", #l, #r, _l, _r, _l, _r, __FILE__, __LINE__); exit(1); }
-#define assertNotEqual(l, r) if ((_l = l) == (_r = r)) { printf("Assertion Failed: assertNotEqual(%s, %s)\n    %llu != %llu (%p == %p)\n# %s:%d\n", #l, #r, _l, _r, _l, _r, __FILE__, __LINE__); exit(1); }
+#define assertEqual(l, r) if ((_l = l) != (_r = r)) { printf("Assertion Failed: assertEqual(%s, %s)\n    %llu != %llu (%p != %p)\n# %s:%d\n", #l, #r, (unsigned long long int) _l, (unsigned long long int) _r, _l, _r, __FILE__, __LINE__); exit(1); }
+#define assertNotEqual(l, r) if ((_l = l) == (_r = r)) { printf("Assertion Failed: assertNotEqual(%s, %s)\n    %llu != %llu (%p == %p)\n# %s:%d\n", #l, #r, (unsigned long long int) _l, (unsigned long long int) _r, _l, _r, __FILE__, __LINE__); exit(1); }
